ticket:211 code used for caching in cart app moved to cache.py

+++ cghub/apps/cart/tests/tests.py
@@ -2,6 +2,7 @@ import os
 import glob
 import shutil
 import datetime
+import shutil
 
 from django.core import mail
 from django.conf import settings
@@ -12,9 +13,11 @@ from django.utils import simplejson as json
 from django.utils import timezone
 from django.utils.importlib import import_module
 from django.contrib.sessions.models import Session
+from django.conf import settings
 
 from cghub.settings.utils import PROJECT_ROOT
 from cghub.apps.cart.forms import SelectedFilesForm, AllFilesForm
+from cghub.apps.cart.cache import save_to_cache, AnalysisFileException
 
 from cghub.apps.core.tests import WithCacheTestCase
 
@@ -209,7 +212,8 @@ class CacheTestCase(TestCase):
         self.api_results_cache_dir = settings.CART_CACHE_DIR
         files = glob.glob(os.path.join(self.api_results_cache_dir, '*'))
         for file in files:
-            os.remove(file)
+            if not os.path.isdir(file):
+                os.remove(file)
         files = glob.glob(os.path.join(testdata_dir, '*'))
         for file in files:
             shutil.copy(file, os.path.join(self.api_results_cache_dir, os.path.basename(file)))
@@ -221,13 +225,45 @@ class CacheTestCase(TestCase):
     def tearDown(self):
         files = glob.glob(os.path.join(self.api_results_cache_dir, '*'))
         for file in files:
-            os.remove(file)
+            if not os.path.isdir(file):
+                pass
+                #os.remove(file)
 
     # test for new cache functions TODO(nanvel): remove this comments and unnecessary tests
-
-
-    # end new tests
-
+    def test_save_to_cache(self):
+        """
+        7b9cd36a-8cbb-4e25-9c08-d62099c15ba1 - 2012-10-29T21:56:12Z
+        """
+        analysis_id = '7b9cd36a-8cbb-4e25-9c08-d62099c15ba1'
+        last_modified = '2012-10-29T21:56:12Z'
+        path = os.path.join(settings.CART_CACHE_DIR, analysis_id)
+        if os.path.exists(path):
+            shutil.rmtree(path)
+        path_full = os.path.join(path, last_modified, 'analysisFull.xml')
+        path_short = os.path.join(path, last_modified, 'analysisShort.xml')
+        save_to_cache(analysis_id, last_modified)
+        self.assertTrue(os.path.exists(path_full))
+        self.assertTrue(os.path.exists(path_short))
+        shutil.rmtree(path)
+        # check exception raises when file does not exists
+        bad_analysis_id = 'bad-analysis-id'
+        path = os.path.join(settings.CART_CACHE_DIR, bad_analysis_id)
+        try:
+            save_to_cache(bad_analysis_id, last_modified)
+        except AnalysisFileException:
+            pass
+        else:
+            raise False, 'AnalysisFileException doesn\'t raised'
+        shutil.rmtree(path)
+        # check case when file was updated
+        path = os.path.join(settings.CART_CACHE_DIR, analysis_id)
+        try:
+            save_to_cache(analysis_id, '1900-10-29T21:56:12Z')
+        except AnalysisFileException:
+            pass
+        else:
+            raise False, 'AnalysisFileException doesn\'t raised'
+        shutil.rmtree(path)
 
 
     def test_cache_generate_manifest_live(self):
